Crate pallet_block_rewards_hybrid

Block Reward Distribution Pallet

Config

Overview

Pallet that implements block reward issuance and distribution mechanics.

After issuing a block reward, pallet will calculate how to distribute the reward based on configurable parameters and chain state.

Major on-chain factors which can influence reward distribution are total issuance and total value locked by dapps staking.

Interface

Dispatchable Function

set_configuration - used to change reward distribution configuration parameters

Other

on_timestamp_set - This pallet implements the OnTimestampSet trait to handle block production. Note: We assume that it’s impossible to set timestamp two times in a block.

Usage

Pallet should be set as a handler of OnTimestampSet.

DappsStakingTvlProvider handler should be defined as an impl of TvlProvider trait. For example:

pub struct TvlProvider();
impl Get<Balance> for TvlProvider {
    fn tvl() -> Balance {
        DappsStaking::total_locked_value()
    }
}

BeneficiaryPayout handler should be defined as an impl of BeneficiaryPayout trait. For example:

pub struct BeneficiaryPayout();
impl BeneficiaryPayout<NegativeImbalanceOf<T>> for BeneficiaryPayout {

    fn treasury(reward: NegativeImbalanceOf<T>) {
        Balances::resolve_creating(&TREASURY_POT.into_account(), reward);
    }

    fn collators(reward: NegativeImbalanceOf<T>) {
        Balances::resolve_creating(&COLLATOR_POT.into_account(), reward);
     }

    fn dapps_staking(stakers: NegativeImbalanceOf<T>, dapps: NegativeImbalanceOf<T>) {
        DappsStaking::rewards(stakers, dapps);
    }
}

Set MaxBlockRewardAmount to the max reward amount distributed per block. Max amount will be reached if ideal_dapps_staking_tvl is reached.

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The pallet module in each FRAME pallet hosts the most important items needed to construct this pallet.

weights

Autogenerated weights for block_rewards_hybrid

Structs

RewardDistributionConfig

List of configuration parameters used to calculate reward distribution portions for all the beneficiaries.

Rewards

Represents rewards distribution balances for each beneficiary

Traits

BeneficiaryPayout

Defines functions used to payout the beneficiaries of block rewards